test guardarEnArray pasado
