+++ Fibo3/salida.txt
@@ -1,0 +1,1 @@
+1 2 3 5 8 13
